From: ericldab@ai.mit.edu (Eric McDonald)
Currently, the extern variable "dst_realm_sz" is defined as a char, when in
fact it is defined as an int in /mit/telnet/src/telnet/libtelnet/kerberos.c.
This causes the "-k" option to fail on this SunOS 4.1.3, and possibly on
other platforms (because the strncpy() copies zero bytes).  Following is a
patch to correct this bug.


git-svn-id: svn+ssh://svn.mit.edu/athena/trunk@7408 728af825-273c-0410-89f9-f7d3b574a069