Promote mitmailutils back to a dependency in clients, dotfiles, xsession.
We had previously dropped it to a recommendation in all three of those
packages to address the fact that it was uninstallable on Karmic and
Squeeze, due to a dependency on libsasl2-krb4-mit.

Now that mitmailutils uses GSSAPI/krb5, the libsasl2-krb4-mit
dependency is gone, and it's installable everywhere.



git-svn-id: svn+ssh://svn.mit.edu/athena/trunk@24307 728af825-273c-0410-89f9-f7d3b574a069