install-debathena.sh: Really don't break Maverick (Trac: #755)
The check needs to happen after we run apt-get update, since the issue
at hand is with a system that all of a sudden tries to process new
updates from Ubuntu.  It also needs to happen before we install
dnsutils, because that's where the first breakage is (see Debian
#610845). So we move it earlier and run apt-get once again before we add
our repo.

However, now that it happens before aptitude is installed, we need to
use apt-get for the check, since that's all we can count on existing.

This updates r24922 "Don't break maverick".



git-svn-id: svn+ssh://svn.mit.edu/athena/trunk@24925 728af825-273c-0410-89f9-f7d3b574a069