Fix dasource when run with a relative path
If you try to dasource using a relative path, instead of using a full path,
it'll error out when trying to run daorig. The reason for this is that daorig
is run as $basedir/daorig, but inside a subshell with a different CWD. Fix this
by using readlink -f to get a full path instead of just initializing $basedir
as $(dirname "$0").


git-svn-id: svn+ssh://svn.mit.edu/athena/trunk@25420 728af825-273c-0410-89f9-f7d3b574a069